requirement.txt
grpcio==1.70.0
grpcio-tools==1.70.0
mysql-connector-python==9.2.0
protobuf==5.29.3
setuptools==75.8.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
